+++ encodings/bytebool/src/lib.rs
@@ -41,11 +41,6 @@
 //! [spec]: https://arrow.apache.org/docs/format/CanonicalExtensions.html#bit-boolean
 
 pub use array::*;
-use vortex_array::ArrayVTable;
-use vortex_array::aggregate_fn::AggregateFnVTable;
-use vortex_array::aggregate_fn::fns::uncompressed_size_in_bytes::FixedWidthUncompressedSizeInBytesKernel;
-use vortex_array::aggregate_fn::fns::uncompressed_size_in_bytes::UncompressedSizeInBytes;
-use vortex_array::aggregate_fn::session::AggregateFnSessionExt;
 use vortex_array::session::ArraySessionExt;
 use vortex_session::VortexSession;
 
@@ -58,9 +53,4 @@ mod slice;
 /// Initialize ByteBool encoding in the given session.
 pub fn initialize(session: &VortexSession) {
     session.arrays().register(ByteBool);
-    session.aggregate_fns().register_aggregate_kernel(
-        ByteBool.id(),
-        Some(UncompressedSizeInBytes.id()),
-        &FixedWidthUncompressedSizeInBytesKernel,
-    );
 }
