+++ encodings/fsst/src/compute/uncompressed_size.rs
@@ -7,6 +7,7 @@ use vortex_array::ArrayRef;
 use vortex_array::ExecutionCtx;
 use vortex_array::aggregate_fn::AggregateFnRef;
 use vortex_array::aggregate_fn::fns::uncompressed_size_in_bytes::UncompressedSizeInBytes;
+use vortex_array::aggregate_fn::fns::uncompressed_size_in_bytes::validity_uncompressed_size_in_bytes;
 use vortex_array::aggregate_fn::kernels::DynAggregateKernel;
 use vortex_array::arrays::PrimitiveArray;
 use vortex_array::arrays::primitive::PrimitiveArrayExt;
@@ -14,7 +15,6 @@ use vortex_array::arrays::varbinview::build_views::BinaryView;
 use vortex_array::dtype::IntegerPType;
 use vortex_array::match_each_integer_ptype;
 use vortex_array::scalar::Scalar;
-use vortex_array::validity::validity_uncompressed_size_in_bytes;
 use vortex_error::VortexResult;
 use vortex_error::vortex_err;
 
@@ -43,12 +43,15 @@ impl DynAggregateKernel for FSSTUncompressedSizeInBytesKernel {
             .map_err(|e| vortex_err!("array length does not fit in u64: {e}"))?
             .checked_mul(size_of::<BinaryView>() as u64)
             .ok_or_else(|| vortex_err!("uncompressed size in bytes overflowed u64"))?;
-        let data_size = uncompressed_lengths_size(
-            &array
+        let data_size = {
+            let lengths: &PrimitiveArray = &array
                 .uncompressed_lengths()
                 .clone()
-                .execute::<PrimitiveArray>(ctx)?,
-        )?;
+                .execute::<PrimitiveArray>(ctx)?;
+            match_each_integer_ptype!(lengths.ptype(), |P| {
+                uncompressed_lengths_size(lengths.as_slice::<P>())
+            })
+        }?;
         let validity_size = validity_uncompressed_size_in_bytes(
             array
                 .as_ref()
@@ -65,13 +68,7 @@ impl DynAggregateKernel for FSSTUncompressedSizeInBytesKernel {
     }
 }
 
-fn uncompressed_lengths_size(lengths: &PrimitiveArray) -> VortexResult<u64> {
-    match_each_integer_ptype!(lengths.ptype(), |P| {
-        uncompressed_lengths_size_typed(lengths.as_slice::<P>())
-    })
-}
-
-fn uncompressed_lengths_size_typed<P: IntegerPType>(lengths: &[P]) -> VortexResult<u64> {
+fn uncompressed_lengths_size<P: IntegerPType>(lengths: &[P]) -> VortexResult<u64> {
     // The lengths child stores decoded byte counts for each logical value.
     let mut size = 0u64;
     for len in lengths {
