+++ encodings/pco/src/lib.rs
@@ -7,22 +7,12 @@ mod rules;
 mod slice;
 
 pub use array::*;
-use vortex_array::ArrayVTable;
-use vortex_array::aggregate_fn::AggregateFnVTable;
-use vortex_array::aggregate_fn::fns::uncompressed_size_in_bytes::FixedWidthUncompressedSizeInBytesKernel;
-use vortex_array::aggregate_fn::fns::uncompressed_size_in_bytes::UncompressedSizeInBytes;
-use vortex_array::aggregate_fn::session::AggregateFnSessionExt;
 use vortex_array::session::ArraySessionExt;
 use vortex_session::VortexSession;
 
 /// Initialize Pco encoding in the given session.
 pub fn initialize(session: &VortexSession) {
     session.arrays().register(Pco);
-    session.aggregate_fns().register_aggregate_kernel(
-        Pco.id(),
-        Some(UncompressedSizeInBytes.id()),
-        &FixedWidthUncompressedSizeInBytesKernel,
-    );
 }
 
 #[derive(Clone, prost::Message)]
