+++ encodings/sequence/src/lib.rs
@@ -20,8 +20,6 @@ use vortex_array::ArrayVTable;
 use vortex_array::aggregate_fn::AggregateFnVTable;
 use vortex_array::aggregate_fn::fns::is_sorted::IsSorted;
 use vortex_array::aggregate_fn::fns::min_max::MinMax;
-use vortex_array::aggregate_fn::fns::uncompressed_size_in_bytes::FixedWidthUncompressedSizeInBytesKernel;
-use vortex_array::aggregate_fn::fns::uncompressed_size_in_bytes::UncompressedSizeInBytes;
 use vortex_array::aggregate_fn::session::AggregateFnSessionExt;
 use vortex_array::session::ArraySessionExt;
 use vortex_session::VortexSession;
@@ -40,11 +38,6 @@ pub fn initialize(session: &VortexSession) {
         Sequence.id(),
         Some(IsSorted.id()),
         &compute::is_sorted::SequenceIsSortedKernel,
-    );
-    session.aggregate_fns().register_aggregate_kernel(
-        Sequence.id(),
-        Some(UncompressedSizeInBytes.id()),
-        &FixedWidthUncompressedSizeInBytesKernel,
     );
 }
