+++ vortex-array/src/aggregate_fn/fns/uncompressed_size_in_bytes/varbinview.rs
@@ -7,9 +7,9 @@ use vortex_error::VortexResult;
 use vortex_error::vortex_err;
 
 use crate::ExecutionCtx;
+use crate::aggregate_fn::fns::uncompressed_size_in_bytes::validity_uncompressed_size_in_bytes;
 use crate::arrays::VarBinViewArray;
 use crate::arrays::varbinview::BinaryView;
-use crate::validity::validity_uncompressed_size_in_bytes;
 
 pub(super) fn varbinview_uncompressed_size_in_bytes(
     array: &VarBinViewArray,
