+++ vortex-array/src/arrays/varbin/compute/uncompressed_size.rs
@@ -10,12 +10,12 @@ use crate::ArrayRef;
 use crate::ExecutionCtx;
 use crate::aggregate_fn::AggregateFnRef;
 use crate::aggregate_fn::fns::uncompressed_size_in_bytes::UncompressedSizeInBytes;
+use crate::aggregate_fn::fns::uncompressed_size_in_bytes::validity_uncompressed_size_in_bytes;
 use crate::aggregate_fn::kernels::DynAggregateKernel;
 use crate::arrays::VarBin;
 use crate::arrays::varbin::VarBinArrayExt;
 use crate::arrays::varbinview::BinaryView;
 use crate::scalar::Scalar;
-use crate::validity::validity_uncompressed_size_in_bytes;
 
 #[derive(Debug)]
 pub(crate) struct VarBinUncompressedSizeInBytesKernel;
