+++ vortex-array/src/validity.rs
@@ -587,23 +587,6 @@ impl IntoArray for &MaskValues {
     }
 }
 
-/// Returns the number of bytes a validity mask contributes to an uncompressed array.
-///
-/// This is the validity portion of the
-/// [`UncompressedSizeInBytes`](crate::aggregate_fn::fns::uncompressed_size_in_bytes::UncompressedSizeInBytes)
-/// aggregate for canonical arrays and encoding kernels that report that aggregate directly.
-///
-/// Uniform masks contribute no bytes, because Vortex does not materialize a validity buffer for
-/// all-valid or all-invalid arrays. A mixed mask contributes the packed bit-buffer size, rounded up
-/// to whole bytes.
-pub fn validity_uncompressed_size_in_bytes(validity: Mask) -> VortexResult<u64> {
-    match validity {
-        Mask::AllTrue(_) | Mask::AllFalse(_) => Ok(0),
-        Mask::Values(values) => u64::try_from(values.len().div_ceil(8))
-            .map_err(|e| vortex_err!("Failed to convert bit buffer length to u64: {e}")),
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use rstest::rstest;
@@ -619,19 +602,6 @@ mod tests {
     use crate::dtype::Nullability;
     use crate::validity::BoolArray;
     use crate::validity::Validity;
-    use crate::validity::validity_uncompressed_size_in_bytes;
-
-    #[rstest]
-    #[case(Mask::AllTrue(9), 0)]
-    #[case(Mask::AllFalse(9), 0)]
-    #[case(Mask::from_iter([true, false, true, true, false, true, false, true, true]), 2)]
-    fn validity_uncompressed_size_matches_validity_buffer_size(
-        #[case] mask: Mask,
-        #[case] expected: u64,
-    ) -> vortex_error::VortexResult<()> {
-        assert_eq!(validity_uncompressed_size_in_bytes(mask)?, expected);
-        Ok(())
-    }
 
     #[rstest]
     #[case(Validity::AllValid, 5, &[2, 4], Validity::AllValid, Validity::AllValid)]
